# src/backend_go/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Cognix Golang API Documentation",
        "title": "Cognix API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/manage/embedding_models": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return list of embedding models",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EmbeddingModel"
                ],
                "summary": "return list of embedding models",
                "operationId": "embedding_model_get_all",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "true for include deleted embedding models",
                        "name": "archived",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.EmbeddingModel"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "creates embedding models",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EmbeddingModel"
                ],
                "summary": "creates embedding models",
                "operationId": "embedding_model_create",
                "parameters": [
                    {
                        "description": "embedding model parameter",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parameters.EmbeddingModelParam"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.EmbeddingModel"
                        }
                    }
                }
            }
        },
        "/manage/embedding_models/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return embedding model by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EmbeddingModel"
                ],
                "summary": "return embedding model by id",
                "operationId": "embedding_model_get_by_id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "embedding model id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.EmbeddingModel"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "updates embedding model",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EmbeddingModel"
                ],
                "summary": "updates embedding model",
                "operationId": "embedding_model_update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "embedding model id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "embedding model parameter",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parameters.EmbeddingModelParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.EmbeddingModel"
                        }
                    }
                }
            }
        },
        "/manage/embedding_models/{id}/{action}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "delete or restore embedding model",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EmbeddingModel"
                ],
                "summary": "delete or restore embedding model",
                "operationId": "embedding_model_delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "embedding model id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "action : delete | restore ",
                        "name": "action",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.EmbeddingModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.EmbeddingModel": {
            "type": "object",
            "properties": {
                "creation_date": {
                    "type": "string"
                },
                "deleted_date": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "id": {
                    "type": "number"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_update": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "model_dim": {
                    "type": "integer"
                },
                "model_id": {
                    "type": "string"
                },
                "model_name": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "parameters.EmbeddingModelParam": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "model_dim": {
                    "type": "integer"
                },
                "model_id": {
                    "type": "string"
                },
                "model_name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "pg.NullTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}